Added New Autotag for Blocks
Blocks can now be displayed with an autotag. Can specify a class which
will wrap it in a div. When a block is displayed via an autotag the
position is ignored (left or right) so those specific block header and
footer template files are ignored. Can still specify own custom header
and footer in theme functions.php file. Part of feature request #683