Added ability for search not to use current topic (feature request #640 )
Added search_use_topic config option. If set to true, when a user visits
the search page (or if the search block is used), the last topic the
user visited will be used as the search topic selection. If set to false
the topic search selection will default to all topics.