Topic Ids and Name Changed from 128 varchar to 75 varchar
Had to reduce the size of the topic ids because the topic assigments
table primary key became to large for tables that use a colliation of
utf8mb4. The original primary key uses 286 (= 128 + 30 + 128)
characters, which will be 858 bytes with utf8 and 1144 bytes with
utf8mb4 which is over the 1000 byte limit.